requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
SQLAlchemy
psycopg2-binary
email-validator